+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added `scroller` prop for limiting scroll distance when `mode` is set to `bottom`, in PR [#XXX](https://github.com/compulim/react-scroll-to-bottom/pull/XXX)
+- Added `initialScrollBehavior` prop for first scroll behavior. When set to `"auto"` (discrete scrolling), it will jump to end on initialization. in PR [#XXX](https://github.com/compulim/react-scroll-to-bottom/pull/XXX)
 
 ### Fixed